+++ tests/benchmark_tests/benchmark_peaks.py
@@ -12,7 +12,7 @@ from harlow.sampling.lola_voronoi import LolaVoronoi
 from harlow.sampling.probabilistic_sampling import Probabilistic_sampler
 from harlow.surrogating.surrogate_model import VanillaGaussianProcess
 from harlow.utils.helper_functions import latin_hypercube_sampling
-from tests.integration_tests.test_functions import peaks_2d
+from harlow.utils.test_functions import peaks_2d
 
 domains_lower_bound = [np.array([-3, -3]), np.array([-5, -5]), np.array([-8, -8])]
 domains_upper_bound = [np.array([3, 3]), np.array([5, 5]), np.array([8, 8])]
