+++ tests/visualization/test_corner.py
@@ -3,8 +3,8 @@ a human."""
 
 import numpy as np
 
+from harlow.utils.test_functions import ackley_nd
 from harlow.visualization.corner import corner
-from tests.integration_tests.test_functions import ackley_nd
 
 
 def func_nd(x: np.ndarray):
